refactor(skins): move shared helpers out of skins/ into lib/
skins/ should only contain skin components; the shared utilities were
general-purpose and belong in lib/:

- sortProjects → lib/projects.ts (next to PROJECTS)
- groupByYear, BlogListData → lib/blog-utils.ts
- formatRelative, getRoleLabel, unlockErrorLabel → lib/format.ts
- PAGE_SIZE → import directly from @blog/shared (was re-exported)
